migrations/requirements.txt
# requirements.txt
alembic==1.13.1
SQLAlchemy==2.0.27
mysqlclient==2.2.4
PyMySQL==1.1.1
python-dotenv==1.0.1
mysql-connector-python==9.1.0